+++ app/Main.hs
@@ -3,4 +3,4 @@ module Main where
 import Ryvm.CLI
 
 main :: IO ()
-main = query
+main = run